Finalised Home page design
