+++ Desktop/pages/data.html
@@ -5,8 +5,8 @@
         <meta charset="utf-8">
         <meta http-equiv="Content-Security-Policy" content="default-src 'self'; script-src 'self'; connect-src http://localhost:3000">
         <title>Integrator - Productivity</title>
-        <script type="module" src="../src/scripts/render/productivity_renderer.js"></script>
-        <link type="text/css" rel="stylesheet" href="../src/styles/productivity.css">
+        <script type="module" src="../src/scripts/render/data_renderer.js"></script>
+        <link type="text/css" rel="stylesheet" href="../src/styles/data.css">
     </head>
     <body>
         <div id="main">
