+++ Desktop/pages/index.html
@@ -14,11 +14,18 @@
                 <img src="../public/img/favicon-32x32.png" alt="therefore logo">
             </nav>                
             <section id="main_area">
-                <header id="title_area"></header>
+                <header id="main_title"></header>
                 <div id="intro_area">
         
                 </div>
-                <div id="selection_area">
+                <!-- <hr> -->
+                <header id="productivity_selection_title"></header>
+                <div id="productivity_selection_area">
+                    
+                </div>
+                <!-- <hr> -->
+                <header id="media_selection_title"></header>
+                <div id="media_selection_area">
                     
                 </div>
             </section>
