+++ Desktop/src/scripts/render/global_renderer.js
@@ -2,33 +2,27 @@ const NAVIGATION_BAR = document.getElementById("navigation_area");
 const NAV_IMAGE = NAVIGATION_BAR.innerHTML;
 
 const Title = (text) => {    
-    return (
-        `
-            <h1>${text}</h1>
-        `
-    );
+    return (`
+       <h1>${text}</h1>
+    `);
 }
 
 
 const Navbar = (props) => {
     let LINKS = "";
-    ["Home", "Productivity Tools", "Media Tools"].map((l) => {
-    return (
-        `
+    ["Home", "Data Tools", "Media Tools"].map((l) => {
+    return (`
         <li>
             <a class="${props.current === l ? "current" : ""}" href="${props.links[l.split(" ")[0]]}">${l}</a>
         </li>
-        `
-        )
+    `)
     }).map((al) => LINKS += al);
     
-    return (
-        `
+    return (`
         <ul>
             ${LINKS}
         </ul>
-        `
-    )
+    `);
 }
 
 export {NAVIGATION_BAR, NAV_IMAGE, Navbar, Title}
