+++ Desktop/src/styles/home.css
@@ -1,10 +1,20 @@
-header#title_area {
+header {
   display: flex;
   justify-content: center;
-  background-color: #fff;
 }
-header#title_area h1 {
-  font-size: 1.25rem;
+
+h1 {
+  font-size: 1.75rem;
+}
+
+h2 {
+  font-size: 1.5em;
+}
+
+#productivity_selection_title, #media_selection_title {
+  margin-top: 0.5rem;
+  margin-bottom: 0rem;
+  padding: 0px;
 }
 
 .current {
@@ -60,6 +70,11 @@ body {
   font-family: Ubuntu;
 }
 
+hr {
+  width: 95%;
+  background-color: #333;
+}
+
 section#main_area {
   height: 100%;
   width: 100%;
@@ -84,4 +99,69 @@ section#main_area div#intro_area ul {
 }
 section#main_area div#intro_area ul ul {
   border: none;
+}
+section#main_area div#productivity_selection_area, section#main_area div#media_selection_area {
+  width: 80%;
+  margin: 0rem auto;
+  padding: 0rem 0.5rem;
+  display: block;
+  justify-content: space-around;
+}
+section#main_area div#productivity_selection_area div.integration-row, section#main_area div#media_selection_area div.integration-row {
+  display: flex;
+  padding: 16px 0px;
+}
+section#main_area div#productivity_selection_area div.integration-row div.integration-section, section#main_area div#media_selection_area div.integration-row div.integration-section {
+  display: block;
+  width: 40%;
+  padding: 8px;
+  font-size: 1rem;
+  height: max-content;
+  margin: auto;
+  border: 1px solid #3f9bef;
+  border-radius: 5px;
+  box-shadow: 5px 5px #3f9bef;
+}
+section#main_area div#productivity_selection_area div.integration-row div.integration-section h2, section#main_area div#media_selection_area div.integration-row div.integration-section h2 {
+  margin: auto;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  font-size: 1.25rem;
+}
+section#main_area div#productivity_selection_area div.integration-row div.integration-section div.image-container, section#main_area div#media_selection_area div.integration-row div.integration-section div.image-container {
+  width: 100%;
+  height: fit-content;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+section#main_area div#productivity_selection_area div.integration-row div.integration-section div.image-container img, section#main_area div#media_selection_area div.integration-row div.integration-section div.image-container img {
+  width: 100px;
+  height: 100px;
+}
+section#main_area div#productivity_selection_area div.integration-row div.integration-section p, section#main_area div#media_selection_area div.integration-row div.integration-section p {
+  display: flex;
+  justify-content: center;
+  padding: 0px;
+  margin-top: 8px;
+  margin-bottom: 0px;
+  font-size: 1em;
+}
+section#main_area div#productivity_selection_area div.integration-row div.integration-section div.button-container, section#main_area div#media_selection_area div.integration-row div.integration-section div.button-container {
+  width: 100%;
+  height: fit-content;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+section#main_area div#productivity_selection_area div.integration-row div.integration-section div.button-container button, section#main_area div#media_selection_area div.integration-row div.integration-section div.button-container button {
+  border: none;
+  background-color: #3f9bef;
+  border-radius: 3px;
+  padding: 8px 16px;
+  font-size: 1rem;
+}
+section#main_area div#productivity_selection_area div.integration-row div.integration-section div.button-container button:hover, section#main_area div#media_selection_area div.integration-row div.integration-section div.button-container button:hover {
+  filter: brightness(125%);
 }
