+++ Desktop/src/styles/scss/home/home.scss
@@ -6,6 +6,11 @@ body {
     font-family: Ubuntu;
 }
 
+hr {
+    width: 95%;
+    background-color: #333;
+}
+
 section#main_area {
     height: 100%;
     width: 100%;
@@ -16,7 +21,7 @@ section#main_area {
         height: max-content;
         width: 70%;
         margin: auto;
-        border: utils.$debug-container;
+        border: 1px solid #000;
         border-radius: 4px;
         box-shadow: 5px 5px #000;
         p {
@@ -28,6 +33,76 @@ section#main_area {
         }
         ul ul {
             border: none;
+        }
+    }
+    div#productivity_selection_area, div#media_selection_area {
+        width: 80%;
+        margin: 0rem auto;
+        padding: 0rem 0.5rem;
+        display: block;
+        justify-content: space-around;
+        // border: utils.$debug-container;
+        div.integration-row {
+            display: flex;
+            // border: utils.$debug-container;
+            padding: 16px 0px;
+            div.integration-section {
+                display: block;
+                width: 40%;
+                padding: 8px;
+                font-size: 1rem;
+                height: max-content;
+                margin: auto;
+                border: 1px solid utils.$main-colour;
+                border-radius: 5px;
+                box-shadow: 5px 5px utils.$main-colour;
+                h2 {
+                    margin: auto;
+                    display: flex;
+                    justify-content: center;
+                    align-items: center;
+                    font-size: 1.25rem;
+                    // border: utils.$debug-container;
+                }
+                div.image-container {
+                    width: 100%;
+                    height: fit-content;
+                    display: flex;
+                    justify-content: center;
+                    align-items: center;
+                    img {
+                        width: 100px;
+                        height: 100px;
+                    }
+                }
+                p {
+                    display: flex;
+                    justify-content: center;
+                    // border: utils.$debug-container;
+                    padding: 0px;
+                    margin-top: 8px;
+                    margin-bottom: 0px;
+                    font-size: 1em;
+                }
+                div.button-container {
+                    width: 100%;
+                    height: fit-content;
+                    display: flex;
+                    justify-content: center;
+                    align-items: center;
+                    button {
+                        border: none;
+                        background-color: utils.$main-colour;
+                        border-radius: 3px;
+                        padding: 8px 16px;
+                        font-size: 1rem;
+                    }
+                    button:hover {
+                        // background-color: adjust-color($color: utils.$main-colour, $alpha: 0.5);
+                        filter: brightness(125%);
+                    }
+                }
+            }
         }
     }
 }
